fix: Доработка режима "Периодический"

+++ src/user/regime/technolog/regime_technolog_periodic.h
@@ -15,15 +15,15 @@ public:
   void processing();
 
 private:
-  float workPeriod_;       //!< Уставка: Время работы
-  float stopPeriod_;       //!< Уставка: Время останова
-  uint32_t workBeginTime_; //!< Время начала работы
-  uint32_t stopBeginTime_; //!< Время начала останова
-  float workTimeToEnd_;    //!< Время до конца работы
-  float stopTimeToEnd_;    //!< Время до конца останова
+  uint32_t workPeriod_;       //!< Уставка: Время работы
+  uint32_t stopPeriod_;       //!< Уставка: Время останова
+  uint32_t workBeginTime_;    //!< Время начала работы
+  uint32_t stopBeginTime_;    //!< Время начала останова
+  float workTimeToEnd_;       //!< Время до конца работы
+  float stopTimeToEnd_;       //!< Время до конца останова
 
-  bool attempt_;           //!< Флаг первой попытки запуска
-  int addTime_;            //!< Добавочное время к паузе
+  bool attempt_;              //!< Флаг первой попытки запуска
+  int addTime_;               //!< Добавочное время к паузе
 
 };
